fix: Phase 2 high-amplification headers — eliminate 57K warnings
Fix conversion warnings in the 11 headers responsible for 67% of total
warning volume. Changes are mechanical casts/suffixes only — no logic changes.

shader.h (ww3d2): static_cast<uint32_t> for enum/mask bitops (28 sites)
w3d_file.h (ww3d2): static_cast<uint8_t> narrowing in RGB ops (23 sites)
rendobj.h (ww3d2): unsigned long Bits → uint32_t, cast enum bitops (12+ sites)
sphere.h (WWMath): static_cast<double/float> in span calculations (28 sites)
wwmath.h (WWMath): f-suffixes on literals, *f() math functions (9 sites)
matrix4.h (WWMath): f-suffixes on all double literals (13 sites)
wwstring.h (wwlib): static_cast<int>(strlen), size_t in memcpy (11 sites)
vector.h (wwlib): size_t in new[], uintptr_t in ID() (9 sites)
widestring.h (wwlib): static_cast<int>(wcslen), size_t in memcpy (12 sites)
winbase.h (compat): static_cast for POSIX call args (4 sites)
bitstream.h (wwbitpack): static_cast<double> for float promotion (3 sites)

Also: lint plugin improvements
- per-type diagnostic messages (char→int8_t, wchar_t→char16_t, etc.)
- char32_t removed from allowed list (now forbidden)
- naked 'unsigned'/'signed' documented as aliases for uint/int
- .clangd: Homebrew LLVM compiler + suppressed diagnostics for plugin files
- Remove #include "global.h" from plugin (LLVM-API-only files)

Warnings: 81,912 → 24,693 (-57,219 / -70%)

+++ original/Code/WWMath/matrix4.h
@@ -205,10 +205,10 @@ WWINLINE Matrix4::Matrix4(const Vector4 & r0, const Vector4 & r1, const Vector4 
  *=============================================================================================*/
 WWINLINE void Matrix4::Make_Identity(void)
 {
-	Row[0].Set(1.0,0.0,0.0,0.0);
-	Row[1].Set(0.0,1.0,0.0,0.0);
-	Row[2].Set(0.0,0.0,1.0,0.0);
-	Row[3].Set(0.0,0.0,0.0,1.0);
+	Row[0].Set(1.0f,0.0f,0.0f,0.0f);
+	Row[1].Set(0.0f,1.0f,0.0f,0.0f);
+	Row[2].Set(0.0f,0.0f,1.0f,0.0f);
+	Row[3].Set(0.0f,0.0f,0.0f,1.0f);
 }
 
 /***********************************************************************************************
@@ -225,7 +225,7 @@ WWINLINE void Matrix4::Make_Identity(void)
  *=============================================================================================*/
 WWINLINE void Matrix4::Init(const Matrix3D & m)
 {
-	Row[0] = m[0]; Row[1] = m[1]; Row[2] = m[2]; Row[3] = Vector4(0.0,0.0,0.0,1.0); 
+	Row[0] = m[0]; Row[1] = m[1]; Row[2] = m[2]; Row[3] = Vector4(0.0f,0.0f,0.0f,1.0f);
 }
 
 /***********************************************************************************************
@@ -323,10 +323,10 @@ WWINLINE void Matrix4::Init_Perspective(float hfov,float vfov,float znear,float 
 	assert(zfar > znear);
 
 	Make_Identity();
-	Row[0][0] = static_cast<float>(1.0 / tan(hfov*0.5));
-	Row[1][1] = static_cast<float>(1.0 / tan(vfov*0.5));
+	Row[0][0] = 1.0f / tanf(hfov * 0.5f);
+	Row[1][1] = 1.0f / tanf(vfov * 0.5f);
 	Row[2][2] = -(zfar + znear) / (zfar - znear);
-	Row[2][3] = static_cast<float>(-(2.0*zfar*znear) / (zfar - znear));
+	Row[2][3] = -(2.0f * zfar * znear) / (zfar - znear);
 	Row[3][2] = -1.0f;
 	Row[3][3] = 0.0f;
 }
@@ -375,12 +375,12 @@ WWINLINE void Matrix4::Init_Perspective
 	assert(zfar > 0.0f);
 
 	Make_Identity();
-	Row[0][0] = static_cast<float>(2.0*znear / (right - left));
+	Row[0][0] = 2.0f * znear / (right - left);
 	Row[0][2] = (right + left) / (right - left);
-	Row[1][1] = static_cast<float>(2.0*znear / (top - bottom));
+	Row[1][1] = 2.0f * znear / (top - bottom);
 	Row[1][2] = (top + bottom) / (top - bottom);
 	Row[2][2] = -(zfar + znear) / (zfar - znear);
-	Row[2][3] = static_cast<float>(-(2.0*zfar*znear) / (zfar - znear));
+	Row[2][3] = -(2.0f * zfar * znear) / (zfar - znear);
 	Row[3][2] = -1.0f;
 	Row[3][3] = 0.0f;
 }
@@ -442,7 +442,7 @@ WWINLINE Matrix4 Matrix4::Inverse() const    // Gauss-Jordan elimination with pa
 		Swap(b.Row[i1], b.Row[j]);
 
 		// Scale row j to have a unit diagonal
-		if (a[j][j]==0.) {
+		if (a[j][j]==0.0f) {
 			//ALGEBRA_ERROR("Matrix4::inverse: singular matrix; can't invert\n");
 		}
 		b.Row[j] /= a.Row[j][j];
